+++ backend/src/config/constants.py
@@ -7,7 +7,7 @@ load_dotenv()
 PROD_ENV = "production"
 DEV_ENV = "development"
 PROJECT_ENV = os.getenv("PROJECT_ENV", DEV_ENV)
-PROJECT_VERSION = "4.0.2"
+PROJECT_VERSION = "5.0.0"
 
 APP_HOST = "0.0.0.0" if PROJECT_ENV == PROD_ENV else "127.0.0.1"
 APP_PORT = 8000