+++ frontend/src/features/countdown/Countdown.tsx
@@ -1,4 +1,5 @@
 const SECONDS_IN_MINUTE = 60;
+const MIN_MINUTES_STRING_LENGTH = 2;
 
 type CountdownProps = {
   isHidden: boolean;
@@ -11,10 +12,9 @@ function Countdown({ isHidden, remainingTime, ...props }: CountdownProps) {
   const remainingMinutes: number = Math.floor(
     remainingTime / SECONDS_IN_MINUTE,
   );
-  let remainingSeconds: string | number = remainingTime % SECONDS_IN_MINUTE;
-  if (remainingSeconds < 10) {
-    remainingSeconds = `0${remainingSeconds}`;
-  }
+  const remainingSeconds: string = String(
+    remainingTime % SECONDS_IN_MINUTE,
+  ).padStart(MIN_MINUTES_STRING_LENGTH, "0");
 
   return (
     <span {...props}>