+++ frontend/src/features/training/index.ts
@@ -1,1 +1,3 @@
+export * from "./olympiadStageStore";
+export * from "./schoolGradeStore";
 export * from "./trainingStore";