+++ frontend/src/pages/matches/MatchesPage.tsx
@@ -1,4 +1,6 @@
+import { GridProvider } from "@/features/grid";
 import { MatchesCountdown } from "./components/MatchesCountdown";
+import { OlympiadStageCheckbox } from "./components/OlympiadStageCheckbox";
 import { SelectGridSize } from "./components/SelectGridSize";
 import { SelectSchoolGrade } from "./components/SelectSchoolGrade";
 import { TrainingArea } from "./components/TrainingArea";
@@ -7,9 +9,12 @@ function MatchesPage() {
   return (
     <>
       <MatchesCountdown />
-      <div className="flex flex-col gap-2 sm:flex-row sm:gap-4">
-        <SelectGridSize />
-        <SelectSchoolGrade />
+      <div className="flex flex-col items-center gap-2 sm:flex-row sm:items-center sm:gap-4">
+        <GridProvider>
+          <SelectSchoolGrade />
+          <SelectGridSize />
+          <OlympiadStageCheckbox />
+        </GridProvider>
       </div>
       <TrainingArea />
     </>